moving distributed circuit outside tensorflow
